+++ libdm/libdm-deptree.c
@@ -141,7 +141,7 @@ static int _dm_snprintf(char *buf, size_t bufsize, const char *format, ...)
         n = vsnprintf(buf, bufsize, format, ap);
         va_end(ap);
 
-        if (n < 0 || (n > bufsize - 1))
+        if (n < 0 || (n > (int) bufsize - 1))
                 return -1;
 
         return n;
